feat(api): Phase API-2 through API-4D — code-first OpenAPI, TS + Python SDKs, real-node validation
- Phase API-2: unified insert DTO, standalone request_id dedup, error codes {error,code}, ingest collection prereq
- Phase API-3 Recovery: Utoipa scaffolding, route manifest tooling, three-way equality verifier
- Phase API-3.1: 74/74 Utoipa coverage, api/openapi/valori-v1.yaml now generator output
- Phase API-3.2: x-status decision, security contract, health/index/ops schema quality, Redocly clean
- Phase API-3.3: response completeness, enum modeling, throwaway SDK validation, contract gate 12/12 PASS
- Phase API-4A: sdk/python + sdk/typescript scaffold, coverage manifests, CI workflows
- Phase API-4C: TypeScript SDK — 74 wrappers, 205 tests, build ESM+CJS+d.ts, real-node validation
- Phase API-4D: Python SDK metadata wire encoding, TS cast elimination, real-node integration both SDKs

+++ crates/valori-effect/src/effect.rs
@@ -59,7 +59,10 @@ pub enum KernelCommandBody {
         values: Vec<f32>,
         text: Option<String>,
         metadata: Option<serde_json::Value>,
-        tag: u8,
+        /// Phase API-2: widened from `u8` to `u64` so the effect-bus command
+        /// can carry the same `tag` the cluster insert path has always
+        /// accepted on the wire, instead of truncating it.
+        tag: u64,
     },
     SoftDeleteRecord {
         record_id: u32,
